fix m11 m12 m13 store

+++ server/api/admin-secure/user-permissions/[id].ts
@@ -1,7 +1,6 @@
 import { getCookie, getRouterParam, createError, readBody } from "h3"
-import type { Customer } from "~/types/customer"
 
-export default defineEventHandler(async (event) => {
+export default defineEventHandler(async (event): Promise<any> => {
   const config = useRuntimeConfig()
   const token = getCookie(event, "token")
   const id = getRouterParam(event, "id")
@@ -9,30 +8,30 @@ export default defineEventHandler(async (event) => {
   if (!id) throw createError({ statusCode: 400, statusMessage: "Missing product ID" })
   if (!token) throw createError({ statusCode: 401, statusMessage: "Unauthorized" })
 
-  const url = `${config.apiBaseUrl}/api/admin-secure/customers/${id}`
+  const url = `${config.apiBaseUrl}/api/admin-secure/user-permissions/${id}`
   const method = event.method
 
 
   try {
 
-    // Show Customer
+    // Show data
     if (method === "GET") {
-      return await $fetch<Customer>(url, {
+      return await $fetch<any>(url, {
         headers: { Authorization: `Bearer ${token}` },
       })
     }
 
-    // Update Customer
-    // if (method === "PUT") {
-    //   const body = await readBody(event)
-    //   return await $fetch(url, {
-    //     method: "PUT",
-    //     body,
-    //     headers: { Authorization: `Bearer ${token}` },
-    //   })
-    // }
+    //Update data
+    if (method === "PUT") {
+      const body = await readBody(event)
+      return await $fetch(url, {
+        method: "PUT",
+        body,
+        headers: { Authorization: `Bearer ${token}` },
+      })
+    }
 
-    // Delete Customer
+    // Delete data
     if (method === "DELETE") {
       return await $fetch<void>(url, {
         method: "DELETE",
